fix pixel rounding bug and change isolation thresh

+++ @iss/extract_and_filter_NoGPU.m
@@ -204,7 +204,7 @@ function o = extract_and_filter_NoGPU(o)
                 
                 if r ~= o.ReferenceRound  
                     %Get histogram data
-                    IFS = int16(IFS);
+                    IFS = int32(IFS);
                     %AbridgedBaseIm = IFS(:,:,8:15);
                     o.HistCounts(:,c,r) = o.HistCounts(:,c,r)+histc(IFS(:),o.HistValues);
                 end
